Change fmt.print os.exit to log.fatal
