+++ .gitignore
@@ -1,0 +1,2 @@
+Dev/
+.DS_Store
