+++ README.md
@@ -1,3 +1,9 @@
 # Get-Gooo
 
 The simple scaffolding CLI tool for Go projects.
+
+Run with
+
+```shell
+go run *.go
+```
