+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/Tmmcmasters/Get-Gooo
 
-go 1.24.3
+go 1.25.1
